requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiomysql==0.2.0
motor==3.3.2
psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.6.1
redis==5.0.1
cassandra-driver==3.28.0
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
psutil==5.9.8
pyyaml==6.0.1
influxdb-client==1.38.0
elasticsearch==8.11.0

